fix(build): Install paru only if not exists

+++ build.sh
@@ -1,10 +1,12 @@
 set -e
 
-sudo pacman -S --needed base-devel
-git clone https://aur.archlinux.org/paru.git
-cd paru
-makepkg -si
-cd .. && rm -r paru
+if [[ -z $(command -v paru) ]]; then
+    sudo pacman -S --needed base-devel
+    git clone https://aur.archlinux.org/paru.git
+    cd paru
+    makepkg -si
+    cd .. && rm -r paru
+fi
 
 paru -S hyprland \
         hyprpaper \
